requirements.txt
asgiref==3.9.1
bleach==6.2.0
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
googletrans==4.0.0rc1
gunicorn==23.0.0
httpx==0.13.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
